Feat: Character Panel UI + Equipping stuff

+++ Assets/Scripts/UIRelated/HandScript.cs
@@ -34,7 +34,10 @@ public class HandScript : MonoBehaviour
     {
         icon.transform.position = Input.mousePosition + offset;
 
-        DeleteItem();
+        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        {
+            DeleteItem();
+        }
     }
 
     public void TakeMoveable(IMoveable moveable)
@@ -61,18 +64,15 @@ public class HandScript : MonoBehaviour
         icon.color = new Color(0, 0, 0, 0);
     }
 
-    private void DeleteItem()
+    public void DeleteItem()
     {
-        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        if (MyMoveable is Item && InventoryScript.MyInstance.FromSlot != null)
         {
-            if (MyMoveable is Item && InventoryScript.MyInstance.FromSlot != null)
-            {
-                (MyMoveable as Item).MySlot.Clear();
-            }
-
-            Drop();
-
-            InventoryScript.MyInstance.FromSlot = null;
+            (MyMoveable as Item).MySlot.Clear();
         }
+
+        Drop();
+
+        InventoryScript.MyInstance.FromSlot = null;
     }
 }
